requirements.txt
pandas
numpy
scikit-learn
python-dotenv
pymongo
pymongo[srv]==3.6
certifi
pyaml



#-e .